Divide model script into model dir, improve doc strings and comments
